test(research): split PDF note region coverage
